+++ app/src/main/java/nextstep/shoppingcart/detail/ProductDetailScreen.kt
@@ -1,6 +1,5 @@
 package nextstep.shoppingcart.detail
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -15,16 +14,12 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.painter.ColorPainter
-import androidx.compose.ui.platform.LocalInspectionMode
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import coil3.compose.AsyncImage
 import nextstep.shoppingcart.R
+import nextstep.shoppingcart.common.ShoppingCartButton
 import nextstep.shoppingcart.common.ProductImage
 import nextstep.shoppingcart.common.ShoppingCartAppBar
 import nextstep.shoppingcart.model.Product
@@ -46,7 +41,10 @@ fun ProductDetailScreen(
             )
         },
         bottomBar = {
-            ProductDetailBottomButton(onClick = onButtonClick)
+            ShoppingCartButton(
+                text = stringResource(R.string.add_to_cart),
+                onClick = onButtonClick
+            )
         }
     ) { innerPadding ->
         Column(
